[CFile] updated to match Kodi Krypton - based on 4fe6cc2, 52e4d3b, a379903

+++ xbmc/filesystem/RTVFile.cpp
@@ -133,7 +133,7 @@ ssize_t CRTVFile::Read(void *lpBuf, size_t uiBufSize)
   if(m_filePos + lenread > m_fileSize)
   {
     CLog::Log(LOGWARNING, "%s - RTV library read passed filesize, returning last chunk", __FUNCTION__);
-    lenread = (m_fileSize - m_filePos);
+    lenread = (ssize_t)(m_fileSize - m_filePos);
     m_filePos = m_fileSize;
     return lenread;
   }
